Arreglado problema con el stock y el carrito

+++ tarea7dwesGeraldSara/src/main/java/com/geraldSara/tarea7dwesGeraldSara/repositorios/EjemplarRepository.java
@@ -45,7 +45,7 @@ public interface EjemplarRepository extends JpaRepository<Ejemplar, Long> {
 	Set<Ejemplar> findByPlantaCodigo(String codigo);
 	
 	 @Query("SELECT COUNT(e) FROM Ejemplar e WHERE e.planta = :planta AND e.disponible = true")
-		Long contarEjemplaresDisponiblesPorPlanta(@Param("planta") Planta planta);
+		int contarEjemplaresDisponiblesPorPlanta(@Param("planta") Planta planta);
 	 
 	 List<Ejemplar> findByPlanta_Id(Long plantaId);
